Shared Migration
Moved what was going to be a shared DLL to XGEF.CoreSystems, including event and network systems.  Converted all projects to use .net standard as a result, since XGEF will now be included in unity.  Overhauled logging objects to not be so dumb.  Added bit flag support functions to enum extensions.  Performed cleanup on everything.